agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
supabase>=2.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
sse-starlette>=2.1.0
boto3>=1.34.0
